separating move harness
