+++ aptos-move/e2e-move-test-harness/Cargo.toml
@@ -1,0 +1,39 @@
+[package]
+name = "aptos-move-e2e-test-harness"
+description = "Move e2e test harness"
+version = "0.1.0"
+
+# Workspace inherited keys
+authors = { workspace = true }
+edition = { workspace = true }
+homepage = { workspace = true }
+license = { workspace = true }
+publish = { workspace = true }
+repository = { workspace = true }
+rust-version = { workspace = true }
+
+[dependencies]
+anyhow = { workspace = true }
+aptos-cached-packages = { workspace = true }
+aptos-framework = { workspace = true }
+aptos-gas-profiling = { workspace = true }
+aptos-gas-schedule = { workspace = true }
+aptos-language-e2e-tests = { workspace = true }
+aptos-rest-client = { workspace = true }
+aptos-transaction-simulation = { workspace = true }
+aptos-types = { workspace = true }
+bcs = { workspace = true }
+claims = { workspace = true }
+move-core-types = { workspace = true }
+move-package = { workspace = true }
+move-symbol-pool = { workspace = true }
+once_cell = { workspace = true }
+project-root = { workspace = true }
+proptest = { workspace = true }
+serde = { workspace = true }
+
+[lib]
+doctest = false
+
+[features]
+testing = ["aptos-gas-schedule/testing"]
