+++ aptos-move/e2e-move-tests/src/aggregator.rs
@@ -1,8 +1,8 @@
 // Copyright (c) Aptos Foundation
 // Licensed pursuant to the Innovation-Enabling Source Code License, available at https://github.com/aptos-labs/aptos-core/blob/main/LICENSE
 
-use crate::{assert_success, harness::MoveHarness};
 use aptos_language_e2e_tests::{account::Account, executor::FakeExecutor};
+use aptos_move_e2e_test_harness::{assert_success, MoveHarness};
 use aptos_types::{account_address::AccountAddress, transaction::SignedTransaction};
 use std::path::PathBuf;
