+++ aptos-move/e2e-move-tests/src/aptos_governance.rs
@@ -1,9 +1,9 @@
 // Copyright (c) Aptos Foundation
 // Licensed pursuant to the Innovation-Enabling Source Code License, available at https://github.com/aptos-labs/aptos-core/blob/main/LICENSE
 
-use crate::harness::MoveHarness;
 use aptos_cached_packages::aptos_stdlib;
 use aptos_language_e2e_tests::account::Account;
+use aptos_move_e2e_test_harness::MoveHarness;
 use aptos_types::{
     account_address::AccountAddress, move_utils::MemberId, transaction::TransactionStatus,
 };
