+++ aptos-move/script-composer/Cargo.toml
@@ -28,8 +28,8 @@ wasm-bindgen = { workspace = true }
 wasm-bindgen-futures = { workspace = true }
 
 [dev-dependencies]
+aptos-move-e2e-test-harness = { workspace = true }
 aptos-types = { workspace = true }
-e2e-move-tests = { path = "../e2e-move-tests" }
 
 [package.metadata.cargo-machete]
 ignored = ["serde_bytes", "wasm-bindgen-futures", "getrandom"]
